requirements.txt
# Core LangChain
langchain==1.2.0
langsmith==0.5.0
langchain-community==0.4.1
langchain-core==1.2.5
langchain-openai==1.1.6
langchain-aws==1.1.0
langgraph==1.0.5
langchain-anthropic==1.3.0
langchain-google-genai==4.1.2

# TTS & Audio
edge-tts>=6.1.0
pydub>=0.25.0
ffmpeg-python>=0.2.0

# Image Processing
Pillow>=10.0.0
httpx>=0.27.0

# Video Processing
moviepy>=1.0.3

# Config & Utils
pyyaml>=6.0
pydantic>=2.0.0

# Optional: OpenAI for DALL-E
openai>=1.0.0
